[api] list layers from sub classes accurately
