+++ api/app/__init__.py
@@ -14,6 +14,7 @@ from app.endpoints import calculation_module, geofile, wms
 from app.healthz import healthz
 from app.models.geofile import create, list_layers
 from app.redirect import redirect_to_api
+from app.common import db
 
 
 def fetch_dataset(base_url, get_parameters, filename, content_type):
@@ -101,6 +102,7 @@ def create_app(environment="production", testing=False):
     app.register_blueprint(api_bp)
     app.register_blueprint(redirect_to_api)
     app.register_blueprint(healthz)
+    app.teardown_appcontext(db.teardown_db)
     with app.app_context():
         if not app.testing:
             init_datasets()
